Relocate more common memory constructs
